+++ bundle.js
@@ -38778,10 +38778,10 @@ module.exports = { // Common-used build variables...
   DEV_DEBUG: DEV_DEBUG,
 
   THEME: "default",
-  buildTag: "v.0.2.8-210128-2031-build-dev-default",
-  timestamp: "2021.01.28, 20:31",
-  timetag: "210128-2031",
-  version: "0.2.8" };
+  buildTag: "v.0.2.9-210129-0030-build-dev-default",
+  timestamp: "2021.01.29, 00:30",
+  timetag: "210129-0030",
+  version: "0.2.9" };
 
 /***/ }),
 
@@ -39064,7 +39064,7 @@ var cssConfig = { // Common-used css variables...
 var formItemSpacing = 4;
 var formItemHeight = 32; // px
 var formItemBorderSize = 1; // px
-var formItemInnerHeight = formItemHeight - formItemBorderSize * 2; // px
+var formItemInnerHeight = formItemHeight; // - formItemBorderSize * 2; // px -- if used css border
 
 Object.assign(cssConfig, { // Form properties...
 
@@ -39556,7 +39556,7 @@ DateTimeSelector = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_
           var date = cbParams[id];
           if (date) {
             var isEndDate = id === 'endDate';
-            var dateObj = _utils_dates__WEBPACK_IMPORTED_MODULE_11__["tuneDateValue"](date, isEndDate, showTime, timeIntervals);
+            var dateObj = _utils_dates__WEBPACK_IMPORTED_MODULE_11__["adjustDateValue"](date, isEndDate, showTime, timeIntervals);
             cbParams[id + 'Obj'] = dateObj;
             cbParams[id] = _utils_dates__WEBPACK_IMPORTED_MODULE_11__["convertDateToType"](dateObj, dateType);
           }
@@ -39935,31 +39935,46 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_3__);
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! react */ "react");
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_4__);
-/* harmony import */ var _utils_configure__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../../utils/configure */ "./src/utils/configure.js");
-/* harmony import */ var _MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ../MenuItem/MenuItem */ "./src/elements/MenuItem/MenuItem.jsx");
-/* harmony import */ var _Menu_pcss__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./Menu.pcss */ "./src/elements/Menu/Menu.pcss");
-/* harmony import */ var _Menu_pcss__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(_Menu_pcss__WEBPACK_IMPORTED_MODULE_7__);
+/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! prop-types */ "./node_modules/prop-types/index.js");
+/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_5__);
+/* harmony import */ var _utils_configure__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ../../utils/configure */ "./src/utils/configure.js");
+/* harmony import */ var _MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ../MenuItem/MenuItem */ "./src/elements/MenuItem/MenuItem.jsx");
+/* harmony import */ var _Menu_pcss__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./Menu.pcss */ "./src/elements/Menu/Menu.pcss");
+/* harmony import */ var _Menu_pcss__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(_Menu_pcss__WEBPACK_IMPORTED_MODULE_8__);
  /** @module Menu
                                                                                                                                                                                                                                                                                 *  @class Menu
                                                                                                                                                                                                                                                                                 *  @since 2020.10.27, 02:58
-                                                                                                                                                                                                                                                                                *  @changed 2020.12.16, 20:21
+                                                                                                                                                                                                                                                                                *  @changed 2021.01.28, 21:55
                                                                                                                                                                                                                                                                                 */
 
 
-// import PropTypes from 'prop-types'
-// import connect from 'react-redux/es/connect/connect'
+
+// import connect from 'react-redux/es/connect/connect';
 
 
 
 
 
 
-var cnMenu = Object(_utils_configure__WEBPACK_IMPORTED_MODULE_5__["cn"])('Menu');
+var cnMenu = Object(_utils_configure__WEBPACK_IMPORTED_MODULE_6__["cn"])('Menu');
 
 // Unique id counter
 var uniqIdCount = 1;var
 
 Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inheritsLoose__WEBPACK_IMPORTED_MODULE_2___default()(Menu, _React$PureComponent);var _proto = Menu.prototype;
+
+
+
+
+
+
+
+
+
+
+
+
+
 
   // Helper fuctions...
   _proto.
@@ -39980,21 +39995,21 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
     return className;
   };_proto.
 
-  setChildrenItemsFromProps = function setChildrenItemsFromProps() /* children, checkedValStates */{var _this2 = this;
+  setChildrenItemsFromProps = function setChildrenItemsFromProps() {var _this2 = this;
     // console.log('Menu:setChildrenItemsFromProps', {
     //   children,
     // })
     var children = this.props.children;
-    var checkedList = [];
+    var selectedList = [];
     if (Array.isArray(children)) {var
       singleChoice = this.props.singleChoice;var _this$props2 =
-      this.props,value = _this$props2.value,checked = _this$props2.checked;
-      var propsChecked = singleChoice && value != null ? [value] : checked;
+      this.props,value = _this$props2.value,selected = _this$props2.selected;
+      var propsSelected = singleChoice && value != null ? [value] : selected;
       children = children.map(function (item) {
         var isArray = !!item && Array.isArray(item);
         var isObject = !!item && typeof item === 'object' && !isArray; // Array.isArray(item)
         var isElement = isObject && /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_4___default.a.isValidElement(item);
-        var isMenuItem = isElement && item.type === _MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_6__["default"];
+        var isMenuItem = isElement && item.type === _MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_7__["default"];
         var isRawObject = isObject && !isElement;
         // console.log('Menu:setChildrenItemsFromProps:item', {
         //   item,
@@ -40007,25 +40022,24 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
           var itemProps = isRawObject ? item : item.props;
           // Construct unique key values...
           var val = itemProps.val;
-          var _checked = Array.isArray(propsChecked) ? propsChecked.includes(val) : itemProps.checked;
-          // const checked = checkedValStates && checkedValStates[val] != null ? checkedValStates[val] : itemProps.checked
-          var checkable = itemProps.checkable != null ? itemProps.checkable : _this2.props.checkable;
+          var checked = Array.isArray(propsSelected) ? propsSelected.includes(val) : itemProps.checked;
+          var checkable = itemProps.checkable != null ? itemProps.checkable : _this2.props.selectable;
           var newProps = _babel_runtime_helpers_extends__WEBPACK_IMPORTED_MODULE_0___default()({},
           itemProps, {
             onClick: itemProps.onClick || _this2.onMenuItemClick,
             checkable: checkable,
-            checked: _checked });
+            checked: checked });
 
           if (isRawObject) {// Raw object -> create MenuItem
             var key = item && item.key || _this2.getId() + '_Item_' + (itemProps.id || itemProps.val);
-            item = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_4___default.a.createElement(_MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_6__["default"], _babel_runtime_helpers_extends__WEBPACK_IMPORTED_MODULE_0___default()({}, newProps, { key: key }));
+            item = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_4___default.a.createElement(_MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_7__["default"], _babel_runtime_helpers_extends__WEBPACK_IMPORTED_MODULE_0___default()({}, newProps, { key: key }));
             // isMenuItem = isElement = true
           } else
           if (isMenuItem) {// MenuItem -> Add onClick handler if handler is not defined
             item = _babel_runtime_helpers_extends__WEBPACK_IMPORTED_MODULE_0___default()({}, item, { props: newProps });
           }
-          if (item.props.checked && (!singleChoice || !checkedList.length)) {
-            checkedList.push(val);
+          if (item.props.checked && (!singleChoice || !selectedList.length)) {
+            selectedList.push(val);
           }
         }
         // TODO: Process arrays (subitems/groups)?
@@ -40034,20 +40048,20 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
     }
     this.setState({
       items: children,
-      checkedList: checkedList });
+      selectedList: selectedList });
 
     // return children
   };_proto.
 
   updateChildrenItems = function updateChildrenItems(checkedValStates) {var _this$props3 =
-    this.props,singleChoice = _this$props3.singleChoice,onChange = _this$props3.onChange;var
+    this.props,id = _this$props3.id,singleChoice = _this$props3.singleChoice,onChange = _this$props3.onChange;var
     items = this.state.items;
-    var checkedList = [];
+    var selectedList = [];
     if (Array.isArray(items)) {
       items = items.map(function (item) {
         var isObject = !!item && typeof item === 'object';
         var isElement = isObject && /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_4___default.a.isValidElement(item);
-        var isMenuItem = isElement && item.type === _MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_6__["default"];
+        var isMenuItem = isElement && item.type === _MenuItem_MenuItem__WEBPACK_IMPORTED_MODULE_7__["default"];
         if (isMenuItem) {
           var itemProps = item.props;var
           val = itemProps.val,checked = itemProps.checked;
@@ -40058,8 +40072,8 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
           if (checkedVal !== checked) {
             item = _babel_runtime_helpers_extends__WEBPACK_IMPORTED_MODULE_0___default()({}, item, { props: _babel_runtime_helpers_extends__WEBPACK_IMPORTED_MODULE_0___default()({}, itemProps, { checked: checkedVal }) });
           }
-          if (checkedVal) {// && (!singleChoice || !checkedList.length)) {
-            checkedList.push(val);
+          if (checkedVal) {// && (!singleChoice || !selectedList.length)) {
+            selectedList.push(val);
           }
         }
         return item;
@@ -40067,12 +40081,12 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
     }
     this.setState({
       items: items,
-      checkedList: checkedList });
+      selectedList: selectedList });
 
     if (typeof onChange === 'function') {
-      var params = { checked: checkedList };
-      if (singleChoice && checkedList.length) {// Add `val` param if singleChoice mode (and has checked)
-        params.value = checkedList[0];
+      var params = { id: id, selected: selectedList };
+      if (singleChoice && selectedList && selectedList.length) {// Add `val` param if singleChoice mode (and has selected)
+        params.value = selectedList[0];
       }
       onChange(params);
     }
@@ -40109,23 +40123,22 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
 
 
 
-
     function (_ref) {var _this$updateChildrenI;var val = _ref.val;var _this$props4 =
       _this.props,onClick = _this$props4.onClick,singleChoice = _this$props4.singleChoice;var
-      checkedList = _this.state.checkedList;
-      var setChecked = !checkedList.includes(val);
-      if (singleChoice === 'forced' && !setChecked) {// Don not made changes if single mode and clicked item was checked
+      selectedList = _this.state.selectedList;
+      var setSelected = !selectedList.includes(val);
+      if (singleChoice === 'forced' && !setSelected) {// Don not made changes if single mode and clicked item was selected
         return;
       }
       if (typeof onClick === 'function') {// Invoke onClick handler
         onClick({ value: val });
       }
-      _this.updateChildrenItems((_this$updateChildrenI = {}, _this$updateChildrenI[val] = setChecked, _this$updateChildrenI)); // Apply items changes
-    });_this.state = {};return _this;}_proto.componentDidMount = function componentDidMount() {// const children = this.props.children
-    this.setChildrenItemsFromProps();};_proto.componentDidUpdate = function componentDidUpdate(prevProps) {var prevChildren = prevProps.children;var children = this.props.children;if (prevChildren !== children) {// TODO: To update items states?
+      _this.updateChildrenItems((_this$updateChildrenI = {}, _this$updateChildrenI[val] = setSelected, _this$updateChildrenI)); // Apply items changes
+    });_this.state = {};return _this;}_proto.componentDidMount = function componentDidMount() {this.setChildrenItemsFromProps();};_proto.componentDidUpdate = function componentDidUpdate(prevProps) {var prevChildren = prevProps.children;var children = this.props.children;if (prevChildren !== children) {// TODO: To update items states?
       this.setChildrenItemsFromProps();}} // Handlers...
   ; // Render...
-  _proto.renderContent = function renderContent() {var
+  _proto.
+  renderContent = function renderContent() {var
     items = this.state.items;
     return items;
   };_proto.
@@ -40135,18 +40148,11 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
 
 
 
-
-
-
-
     this.props,id = _this$props5.id,setDomRef = _this$props5.setDomRef;
 
     var renderProps = {
       id: id,
       className: this.getClassName(),
-      // title,
-      // onClick,
-      // key,
       ref: setDomRef // Init ref for FormItemHOC
     };
 
@@ -40157,8 +40163,8 @@ Menu = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inhe
       content));
 
 
-  };return Menu;}(react__WEBPACK_IMPORTED_MODULE_4___default.a.PureComponent /** @lends @Menu.prototype */);
-
+  };return Menu;}(react__WEBPACK_IMPORTED_MODULE_4___default.a.PureComponent /** @lends @Menu.prototype */);_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_3___default()(Menu, "propTypes", { className: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.string, disabled: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.bool, id: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.string, layout: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.string, mode: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.string, // ???
+  onChange: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.func, onClick: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.func, selectable: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.bool, setDomRef: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.func, singleChoice: prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_5___default.a.bool]) });
 
 
 /* harmony default export */ __webpack_exports__["default"] = (Menu);
@@ -40193,11 +40199,13 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2__);
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! react */ "react");
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_3__);
-/* harmony import */ var _utils_configure__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../utils/configure */ "./src/utils/configure.js");
-/* harmony import */ var _InlineIcon_InlineIcon__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../InlineIcon/InlineIcon */ "./src/elements/InlineIcon/InlineIcon.jsx");
-/* harmony import */ var _forms_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ../../forms/FormItemHOC/FormItemHOC */ "./src/forms/FormItemHOC/FormItemHOC.jsx");
-/* harmony import */ var _MenuItem_pcss__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./MenuItem.pcss */ "./src/elements/MenuItem/MenuItem.pcss");
-/* harmony import */ var _MenuItem_pcss__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(_MenuItem_pcss__WEBPACK_IMPORTED_MODULE_7__);
+/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! prop-types */ "./node_modules/prop-types/index.js");
+/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_4__);
+/* harmony import */ var _utils_configure__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../../utils/configure */ "./src/utils/configure.js");
+/* harmony import */ var _InlineIcon_InlineIcon__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ../InlineIcon/InlineIcon */ "./src/elements/InlineIcon/InlineIcon.jsx");
+/* harmony import */ var _forms_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ../../forms/FormItemHOC/FormItemHOC */ "./src/forms/FormItemHOC/FormItemHOC.jsx");
+/* harmony import */ var _MenuItem_pcss__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./MenuItem.pcss */ "./src/elements/MenuItem/MenuItem.pcss");
+/* harmony import */ var _MenuItem_pcss__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(_MenuItem_pcss__WEBPACK_IMPORTED_MODULE_8__);
  /** @module MenuItem
                                                                                                                                                                                                                           *  @class MenuItem
                                                                                                                                                                                                                           *  @since 2020.10.27, 03:05
@@ -40205,8 +40213,8 @@ __webpack_require__.r(__webpack_exports__);
                                                                                                                                                                                                                           */
 
 
-// import PropTypes from 'prop-types'
-// import connect from 'react-redux/es/connect/connect'
+
+// import connect from 'react-redux/es/connect/connect';
 
 
 
@@ -40215,9 +40223,24 @@ __webpack_require__.r(__webpack_exports__);
 
 
 
-var cnMenuItem = Object(_utils_configure__WEBPACK_IMPORTED_MODULE_4__["cn"])('MenuItem');var
+var cnMenuItem = Object(_utils_configure__WEBPACK_IMPORTED_MODULE_5__["cn"])('MenuItem');var
 
 MenuItem = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inheritsLoose__WEBPACK_IMPORTED_MODULE_1___default()(MenuItem, _React$PureComponent);function MenuItem() {var _this;for (var _len = arguments.length, args = new Array(_len), _key = 0; _key < _len; _key++) {args[_key] = arguments[_key];}_this = _React$PureComponent.call.apply(_React$PureComponent, [this].concat(args)) || this;_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this), "onClick",
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 
@@ -40247,6 +40270,7 @@ MenuItem = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_
       checked: checked,
       disabled: disabled },
     [this.props.className]);
+    // console.log('MenuItem:constructor', this.props);
     return className;
   }
 
@@ -40261,7 +40285,7 @@ MenuItem = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_
     this.props,withIcon = _this$props3.withIcon,checkable = _this$props3.checkable,checked = _this$props3.checked,icon = _this$props3.icon;
     if (withIcon && icon || checkable && checked) {
       var iconContent = icon || 'faCheck';
-      return iconContent && /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_3___default.a.createElement(_InlineIcon_InlineIcon__WEBPACK_IMPORTED_MODULE_5__["default"], { icon: iconContent, className: cnMenuItem('Icon') });
+      return iconContent && /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_3___default.a.createElement(_InlineIcon_InlineIcon__WEBPACK_IMPORTED_MODULE_6__["default"], { icon: iconContent, className: cnMenuItem('Icon') });
     }
   };_proto.
 
@@ -40303,11 +40327,11 @@ MenuItem = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_
       textContent));
 
 
-  };return MenuItem;}(react__WEBPACK_IMPORTED_MODULE_3___default.a.PureComponent /** @lends @MenuItem.prototype */);
+  };return MenuItem;}(react__WEBPACK_IMPORTED_MODULE_3___default.a.PureComponent /** @lends @MenuItem.prototype */);_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(MenuItem, "propTypes", { checkable: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, checked: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, className: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, disabled: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, htmlId: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, // ???
+  icon: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.object]), id: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, onClick: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.func, setDomRef: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.func, // From FormItemHOC
+  text: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, val: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number]), withIcon: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool });
 
-
-
-/* harmony default export */ __webpack_exports__["default"] = (Object(_forms_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_6__["default"])({ solid: true, hoverable: true })(MenuItem));
+/* harmony default export */ __webpack_exports__["default"] = (Object(_forms_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_7__["default"])({ solid: true, hoverable: true })(MenuItem));
 
 /***/ }),
 
@@ -40967,6 +40991,7 @@ ModalPopup = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
       if (nextOpen) {
         if (_this.controlNode) {// Move focus out of control element...
           // this.controlNode.focus && this.controlNode.focus(); // Ensure focus isnt on other element (eg, selects' inner button).
+          // console.log('ModalPopup:onControlClick: blur');
           _this.controlNode.blur && _this.controlNode.blur();
         }
         if (_this.controlDomNode) {// Move focus out of control element...
@@ -42831,8 +42856,10 @@ __webpack_require__.r(__webpack_exports__);
 
 
 
-var cnFormButton = Object(_utils_configure__WEBPACK_IMPORTED_MODULE_7__["cn"])('FormButton');var
+var cnFormButton = Object(_utils_configure__WEBPACK_IMPORTED_MODULE_7__["cn"])('FormButton');
 
+var defaultTag = 'button'; // 'div'; // `button` element is 'labelable', @see `https://stackoverflow.com/questions/54792126/html-label-with-for-div-id-to-focus-a-contenteditable-div-does-label-label`
+var
 FormButton = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helpers_inheritsLoose__WEBPACK_IMPORTED_MODULE_2___default()(FormButton, _React$PureComponent);
 
 
@@ -42943,14 +42970,12 @@ FormButton = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
     function () {var
       formItemNode = _this.props.formItemNode;
       // console.log('FormButton:focus');
-      // debugger;
       formItemNode && formItemNode.focus && formItemNode.focus();
     });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_3___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_1___default()(_this), "blur",
     function () {var
       formItemNode = _this.props.formItemNode;
-      // console.log('FormButton:focus');
-      // debugger;
-      formItemNode && formItemNode.focus && formItemNode.focus();
+      // console.log('FormButton:blur');
+      formItemNode && formItemNode.blur && formItemNode.blur();
     });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_3___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_1___default()(_this), "onClick",
 
     function (event) {var _this$props =
@@ -43023,7 +43048,8 @@ FormButton = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
 
 
 
-    this.props,id = _this$props6.id,disabled = _this$props6.disabled,tag = _this$props6.tag,type = _this$props6.type,title = _this$props6.title,setDomRef = _this$props6.setDomRef;
+
+    this.props,id = _this$props6.id,inputId = _this$props6.inputId,disabled = _this$props6.disabled,tag = _this$props6.tag,type = _this$props6.type,title = _this$props6.title,setDomRef = _this$props6.setDomRef;
 
     var iconElem = this.renderIcon(); // Icon element
     var textElem = this.renderText(); // Text element
@@ -43031,14 +43057,14 @@ FormButton = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
     // const basicRenderProps = this.getRenderProps() // Get from props not from overrided `super`
     var renderProps = {
       // ...basicRenderProps,
-      id: id,
+      id: inputId || id,
       className: this.getClassName(),
       disabled: disabled,
       title: title,
       type: type,
       onClick: this.onClick,
       ref: setDomRef, // Init ref for FormItemHOC
-      tabIndex: -1 };
+      tabIndex: 0 };
 
 
     var content = /*#__PURE__*/
@@ -43048,7 +43074,7 @@ FormButton = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
 
 
 
-    var tagName = tag || 'div';
+    var tagName = tag || defaultTag;
     var element = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_4___default.a.createElement(tagName, renderProps, content);
     return element;
   };return FormButton;}(react__WEBPACK_IMPORTED_MODULE_4___default.a.PureComponent /** @lends @FormButton.prototype */);
@@ -43072,7 +43098,7 @@ _babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_3___default()(For
   theme: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.string // Button style (plain, default, primary, secondary, error, warn, success, info, etc -- some are in progress -- see styles file)
   // setDomRef: PropTypes.func,
   // setNodeRef: PropTypes.func,
-});/* harmony default export */ __webpack_exports__["default"] = (Object(redux__WEBPACK_IMPORTED_MODULE_5__["compose"])(_helpers_ActionsContext__WEBPACK_IMPORTED_MODULE_8__["withActionsContext"], Object(_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_10__["default"])({ solid: true, hoverable: true, framed: true }))(FormButton));
+});/* harmony default export */ __webpack_exports__["default"] = (Object(redux__WEBPACK_IMPORTED_MODULE_5__["compose"])(_helpers_ActionsContext__WEBPACK_IMPORTED_MODULE_8__["withActionsContext"], Object(_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_10__["default"])({ solid: true, hoverable: true, focusable: true, framed: true }))(FormButton));
 
 /***/ }),
 
@@ -43416,7 +43442,9 @@ FormDateRange = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_hel
         text: buttonText,
         title: title,
         fullWidth: fullWidth,
-        disabled: disabled }));
+        disabled: disabled,
+        checkable: true,
+        checked: open }));
 
 
   };_proto.
@@ -44497,7 +44525,7 @@ __webpack_require__.r(__webpack_exports__);
  /** @module FormItemHOC
                                                                                                                                                                                                                                                                                 *  @class FormItemHOC
                                                                                                                                                                                                                                                                                 *  @since 2019.09.25, 19:18
-                                                                                                                                                                                                                                                                                *  @changed 2020.06.04, 22:58
+                                                                                                                                                                                                                                                                                *  @changed 2021.01.28, 21:58
                                                                                                                                                                                                                                                                                 */
 
 
@@ -44513,7 +44541,7 @@ var classNameModifiers = [// Pass props/state params to class modifiers
 'fullWidth',
 // 'innerFlex', // (???)
 // 'solid', // Item is unwrappable (???)
-// 'framed', // Frame around item (input field, select etc) (???)
+'framed', // Frame around item (input field, button, select etc)
 // Behavior-related modifiers...
 'hoverable', // Allow hover effects
 'focusable', // Allow hover effects
@@ -44569,6 +44597,7 @@ var deriveState = function deriveState() {for (var _len = arguments.length, sets
 var uniqIdCount = 1;
 
 var wrapFormItemHOC = function wrapFormItemHOC(WrappedComponent, params) {var _class, _temp;if (params === void 0) {params = {};}return _temp = _class = /*#__PURE__*/function (_React$Component) {_babel_runtime_helpers_inheritsLoose__WEBPACK_IMPORTED_MODULE_1___default()(FormItem, _React$Component);var _proto = FormItem.prototype;
+
 
 
 
@@ -44681,19 +44710,17 @@ var wrapFormItemHOC = function wrapFormItemHOC(WrappedComponent, params) {var _c
 
 
 
-
-
       function () {var _assertThisInitialize = _babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this),
         formItemDomRef = _assertThisInitialize.formItemDomRef;
         // console.log('FormItem:focus');
-        // debugger;
         formItemDomRef && formItemDomRef.focus && formItemDomRef.focus();
+        _this.handleFocusIn();
       });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this), "blur",
       function () {var _assertThisInitialize2 = _babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this),
         formItemDomRef = _assertThisInitialize2.formItemDomRef;
-        // console.log('FormItem:focus');
-        // debugger;
-        formItemDomRef && formItemDomRef.focus && formItemDomRef.focus();
+        // console.log('FormItem:blur');
+        formItemDomRef && formItemDomRef.blur && formItemDomRef.blur();
+        _this.handleFocusOut();
       });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this), "handleMouseOver",
 
       function () {
@@ -44738,11 +44765,9 @@ var wrapFormItemHOC = function wrapFormItemHOC(WrappedComponent, params) {var _c
         }
         // domRef && domRef.focus && domRef.focus() // ???
       });_this.state = deriveState(defaultState, params, props); // deriveStateFromProps(props, defaultState)
-      _this.id = props.id || params.id; // this.formItemRef = React.createRef()
-      return _this;}FormItem.getDerivedStateFromProps = function getDerivedStateFromProps(props, state) {// TODO: Update event subscriptions if `hoverable` flag changed?
+      _this.id = props.id || params.id;return _this;}FormItem.getDerivedStateFromProps = function getDerivedStateFromProps(props, state) {// TODO: Update event subscriptions if `hoverable` flag changed?
       return deriveState(params, props, state); // deriveStateFromProps(props, state)
-    };_proto.componentDidMount = function componentDidMount() {// const { formItemRef: { current } = {} } = this
-      var formItemDomRef = this.formItemDomRef;if (formItemDomRef && formItemDomRef.addEventListener) {var hoverable = this.getStateOrPropOrParam('hoverable');var focusable = this.getStateOrPropOrParam('focusable');if (hoverable && !this.hoverableInited) {this.hoverableInited = true;formItemDomRef.addEventListener('mouseover', this.handleMouseOver);formItemDomRef.addEventListener('mouseout', this.handleMouseOut);}if (focusable && !this.focusableInited) {this.focusableInited = true; // formItemDomRef.addEventListener('focus', this.handleFocusIn);
+    };_proto.componentDidMount = function componentDidMount() {var formItemDomRef = this.formItemDomRef;if (formItemDomRef && formItemDomRef.addEventListener) {var hoverable = this.getStateOrPropOrParam('hoverable');var focusable = this.getStateOrPropOrParam('focusable');if (hoverable && !this.hoverableInited) {this.hoverableInited = true;formItemDomRef.addEventListener('mouseover', this.handleMouseOver);formItemDomRef.addEventListener('mouseout', this.handleMouseOut);}if (focusable && !this.focusableInited) {this.focusableInited = true; // formItemDomRef.addEventListener('focus', this.handleFocusIn);
           formItemDomRef.addEventListener('focus', this.handleFocusIn);formItemDomRef.addEventListener('blur', this.handleFocusOut);formItemDomRef.addEventListener('focusin', this.handleFocusIn);formItemDomRef.addEventListener('focusout', this.handleFocusOut); // console.log('FormItemHOC:componentDidMount: focus event handlers added');
           this.focus = function () {// Focus handler
             formItemDomRef && formItemDomRef.focus && formItemDomRef.focus();};}}var setFormItemNodeRef = this.props.setFormItemNodeRef;if (typeof setFormItemNodeRef === 'function') {setFormItemNodeRef(this);}};_proto.componentWillUnmount = function componentWillUnmount() {var formItemDomRef = this.formItemDomRef;if (formItemDomRef && formItemDomRef.removeEventListener) {// const hoverable = this.getStateOrPropOrParam('hoverable')
@@ -44750,9 +44775,11 @@ var wrapFormItemHOC = function wrapFormItemHOC(WrappedComponent, params) {var _c
         if (this.hoverableInited) {formItemDomRef.removeEventListener('mouseover', this.handleMouseOver);formItemDomRef.removeEventListener('mouseout', this.handleMouseOut);}if (this.focusableInited) {formItemDomRef.removeEventListener('focus', this.handleFocusIn);formItemDomRef.removeEventListener('blur', this.handleFocusOut);formItemDomRef.removeEventListener('focusin', this.handleFocusIn);formItemDomRef.removeEventListener('focusout', this.handleFocusOut);}}};_proto.getClassName = function getClassName() {var _this2 = this; // Collect modifier values from state or props
       var mods = classNameModifiers.reduce(function (mods, id) {var val = _this2.getStateOrPropOrParam(id); // (this.state[id] != null) ? this.state[id] : this.props[id]
         if (val != null) {mods[id] = val;}return mods;}, {});return cnFormItem(mods, [this.props.className]);} // Handlers...
-    ;_proto.render = function render() {var _this$state = this.state,hovered = _this$state.hovered,focused = _this$state.focused;var renderProps = { hovered: hovered, focused: focused, formItemNode: this, className: this.getClassName(), setDomRef: this.setDomRef // Children dom node receiver
+    ;_proto.render = function render() {var _this$state = this.state,hovered = _this$state.hovered,focused = _this$state.focused;var renderProps = { hovered: hovered, focused: focused, formItemNode: this, className: this.getClassName(),
+        setDomRef: this.setDomRef // Children dom node receiver
         // formItemDomRef: this.formItemDomRef,
-      };var focusable = this.getStateOrPropOrParam('focusable');
+      };
+      var focusable = this.getStateOrPropOrParam('focusable');
       if (focusable) {
         renderProps.tabIndex = 0;
       }
@@ -44762,14 +44789,14 @@ var wrapFormItemHOC = function wrapFormItemHOC(WrappedComponent, params) {var _c
         renderProps)));
 
 
-    };return FormItem;}(react__WEBPACK_IMPORTED_MODULE_4___default.a.Component), _babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_class, "propTypes", { checked: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, className: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.string, clickable: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, disabled: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, focusable: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, focused: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, fullWidth: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, hoverable: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, hovered: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, id: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.string, pressed: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, setDomRef: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.func, setFormItemNodeRef: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.func }), _temp;};
+    };return FormItem;}(react__WEBPACK_IMPORTED_MODULE_4___default.a.Component), _babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_class, "propTypes", { checked: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, className: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.string, clickable: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, disabled: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, focusable: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, focused: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, framed: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, fullWidth: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, hoverable: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, hovered: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, id: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.string, pressed: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.bool, setDomRef: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.func, setFormItemNodeRef: prop_types__WEBPACK_IMPORTED_MODULE_6___default.a.func }), _temp;};
 
 
 
 /** Usage:
-                                                                                                                                                                                                                                                                                                                                                                                                                                                          * FormItemHOC(component)
-                                                                                                                                                                                                                                                                                                                                                                                                                                                          * FormItemHOC(params)(component)
-                                                                                                                                                                                                                                                                                                                                                                                                                                                          */
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  * FormItemHOC(component)
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  * FormItemHOC(params)(component)
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  */
 var FormItemHOC = function FormItemHOC(params) {
   if (typeof params === 'string') {// Passed identifier
     params = { id: params };
@@ -45445,15 +45472,15 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
     });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this), "onMenuChange",
     function (params) {var
       onChange = _this.props.onChange;var
-      checked = params.checked;
+      selected = params.selected;
       if (typeof onChange === 'function') {var _this$props2 =
         _this.props,id = _this$props2.id,inputId = _this$props2.inputId,name = _this$props2.name,singleChoice = _this$props2.singleChoice;
-        var value = singleChoice ? checked[0] : checked;
+        var value = singleChoice ? selected[0] : selected;
         var setId = id || inputId || name;
-        var setParams = { id: setId, checked: checked, value: value };
+        var setParams = { id: setId, selected: selected, value: value };
         onChange(setParams);
       }
-      _this.setState({ checked: checked });
+      _this.setState({ selected: selected });
     });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this), "setPopupRef",
 
     function (node) {var
@@ -45462,11 +45489,15 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
       if (setPopupNodeRef && typeof setPopupNodeRef === 'function') {
         setPopupNodeRef(node);
       }
-    });_this.id = props.id || props.inputId || props.name;var _checked = props.checked,_value = props.value;_this.state = { checked: Array.isArray(_checked) ? _checked : _value && [_value] || [] }; // if (props.setNodeRef) {
+    });_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(_babel_runtime_helpers_assertThisInitialized__WEBPACK_IMPORTED_MODULE_0___default()(_this), "handleOpenState",
+
+    function (_ref) {var open = _ref.open;
+      _this.setState({ open: open });
+    });_this.id = props.id || props.inputId || props.name;var _selected = props.selected,_value = props.value;_this.state = { selected: Array.isArray(_selected) ? _selected : _value && [_value] || [] }; // if (props.setNodeRef) {
     //   props.setNodeRef(this);
     // }
     return _this;} // Helper methods...
-  var _proto = FormSelect.prototype;_proto.getClassName = function getClassName() {var id = this.id;var classList = cnFormSelect({ id: id }, [this.props.className]);return classList;};_proto.getItemsText = function getItemsText() {var checked = this.state.checked;var options = this.props.options;var text = Array.isArray(options) && Array.isArray(checked) && options.map(function (_ref) {var val = _ref.val,text = _ref.text;if (checked.includes(val)) {return text;}}).filter(Boolean).join(', ');return text;} // Handlers...
+  var _proto = FormSelect.prototype;_proto.getClassName = function getClassName() {var id = this.id;var classList = cnFormSelect({ id: id }, [this.props.className]);return classList;};_proto.getItemsText = function getItemsText() {var selected = this.state.selected;var options = this.props.options;var text = Array.isArray(options) && Array.isArray(selected) && options.map(function (_ref2) {var val = _ref2.val,text = _ref2.text;if (selected.includes(val)) {return text;}}).filter(Boolean).join(', ');return text;} // Handlers...
   ; // Render...
   _proto.renderControlContent = function renderControlContent() {var _this$props3 =
 
@@ -45474,10 +45505,15 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
 
 
 
-    this.props,text = _this$props3.text,placeholder = _this$props3.placeholder,title = _this$props3.title,controlButtonTheme = _this$props3.controlButtonTheme,_this$props3$fullWidt = _this$props3.fullWidth,fullWidth = _this$props3$fullWidt === void 0 ? true : _this$props3$fullWidt,disabled = _this$props3.disabled;
+
+    this.props,text = _this$props3.text,placeholder = _this$props3.placeholder,title = _this$props3.title,controlButtonTheme = _this$props3.controlButtonTheme,_this$props3$fullWidt = _this$props3.fullWidth,fullWidth = _this$props3$fullWidt === void 0 ? true : _this$props3$fullWidt,disabled = _this$props3.disabled,inputId = _this$props3.inputId;var
+
+    open =
+    this.state.open;
     var buttonText = this.getItemsText() || placeholder || text;
     return /*#__PURE__*/(
       react__WEBPACK_IMPORTED_MODULE_3___default.a.createElement(_FormButton_FormButton__WEBPACK_IMPORTED_MODULE_9__["default"], {
+        inputId: inputId,
         icon: "faChevronDown",
         rightIcon: true,
         theme: controlButtonTheme || 'primary',
@@ -45486,7 +45522,9 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
         text: buttonText,
         title: title,
         fullWidth: fullWidth,
-        disabled: disabled }));
+        disabled: disabled,
+        checkable: true,
+        checked: open }));
 
 
   };_proto.
@@ -45497,19 +45535,17 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
 
 
 
-
-
     this.props,singleChoice = _this$props4.singleChoice,options = _this$props4.options,disabled = _this$props4.disabled;var
 
-    checked =
-    this.state.checked;
+    selected =
+    this.state.selected;
     return /*#__PURE__*/(
       react__WEBPACK_IMPORTED_MODULE_3___default.a.createElement(_elements_Menu_Menu__WEBPACK_IMPORTED_MODULE_8__["default"], {
-        checkable: true,
+        selectable: true,
         singleChoice: singleChoice,
         onChange: this.onMenuChange,
         onClick: this.onMenuItemClick,
-        checked: checked
+        selected: selected
         // value={value}
         , disabled: disabled },
 
@@ -45544,7 +45580,8 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
       onControlClick: this.onControlClick,
       fullWidth: fullWidth,
       ref: this.setPopupRef,
-      setDomRef: setDomRef };
+      setDomRef: setDomRef,
+      handleOpenState: this.handleOpenState };
 
 
     return /*#__PURE__*/(
@@ -45552,7 +45589,7 @@ FormSelect = /*#__PURE__*/function (_React$PureComponent) {_babel_runtime_helper
 
 
   };return FormSelect;}(react__WEBPACK_IMPORTED_MODULE_3___default.a.PureComponent /** @lends @FormSelect.prototype */);_babel_runtime_helpers_defineProperty__WEBPACK_IMPORTED_MODULE_2___default()(FormSelect, "propTypes", { // value: PropTypes.oneOfType([ PropTypes.string, PropTypes.number, PropTypes.arrayOf(PropTypes.oneOfType([ PropTypes.string, PropTypes.number ])) ]),
-  checked: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.arrayOf(prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number])), disabled: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, fullWidth: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, id: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, onChange: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.func, open: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, options: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.arrayOf(prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.shape({ val: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number]), text: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string })), placeholder: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, text: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, value: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number]), setDomRef: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.func // setNodeRef: PropTypes.func,
+  selected: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.arrayOf(prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number])), disabled: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, fullWidth: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, id: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, onChange: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.func, open: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.bool, options: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.arrayOf(prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.shape({ val: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number]), text: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string })), placeholder: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, text: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, value: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.oneOfType([prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.string, prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.number]), setDomRef: prop_types__WEBPACK_IMPORTED_MODULE_4___default.a.func // setNodeRef: PropTypes.func,
 });
 
 /* harmony default export */ __webpack_exports__["default"] = (Object(_FormItemHOC_FormItemHOC__WEBPACK_IMPORTED_MODULE_6__["default"])({ solid: true, hoverable: true })(FormSelect));
@@ -46346,18 +46383,20 @@ var cn = function cn() {for (var _len = arguments.length, args = new Array(_len)
 /*!****************************!*\
   !*** ./src/utils/dates.js ***!
   \****************************/
-/*! exports provided: detectDateValueType, convertToDateObject, toMicrosoftDateTime, convertDateToType, formatDateToString, formatDateTimeToString, tuneDateValue, startOfTheDay, endOfTheDay */
+/*! exports provided: dayTicks, weekTicks, detectDateValueType, convertToDateObject, toMicrosoftDateTime, convertDateToType, formatDateToString, formatDateTimeToString, adjustDateValue, startOfTheDay, endOfTheDay */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "dayTicks", function() { return dayTicks; });
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "weekTicks", function() { return weekTicks; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "detectDateValueType", function() { return detectDateValueType; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "convertToDateObject", function() { return convertToDateObject; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "toMicrosoftDateTime", function() { return toMicrosoftDateTime; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "convertDateToType", function() { return convertDateToType; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "formatDateToString", function() { return formatDateToString; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "formatDateTimeToString", function() { return formatDateTimeToString; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "tuneDateValue", function() { return tuneDateValue; });
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "adjustDateValue", function() { return adjustDateValue; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "startOfTheDay", function() { return startOfTheDay; });
 /* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "endOfTheDay", function() { return endOfTheDay; });
 /* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! moment */ "moment");
@@ -46376,11 +46415,16 @@ __webpack_require__.r(__webpack_exports__);
 
 
 
+// Export constants...
+
+var dayTicks = 1000 * 60 * 60 * 24;
+var weekTicks = dayTicks * 7;
+
 /** Get date type
-                                        * @param {Date|Moment|number} date
-                                        * @return {String} dateType ('number', 'object', 'moment', 'msDateStr')
-                                        * msDateStr: '/Date(1498653276417+0300)/'
-                                        */
+                                      * @param {Date|Moment|number} date
+                                      * @return {String} dateType ('number', 'object', 'moment', 'msDateStr')
+                                      * msDateStr: '/Date(1498653276417+0300)/'
+                                      */
 function detectDateValueType(date) {
   if (date == null) {
     return undefined;
@@ -46477,14 +46521,14 @@ function formatDateTimeToString(date, opt) {if (opt === void 0) {opt = {};}
   return Object(date_fns__WEBPACK_IMPORTED_MODULE_1__["format"])(convertToDateObject(date), fmt, opt);
 }
 
-/** tuneDateValue -- Adjust date to day/interval start/end.
+/** adjustDateValue -- Adjust date to day/interval start/end.
    * @param {Date|Number|moment|msDateStr} origDate - date (in any format)
    * @param {Boolean} [toEnd=false] -- Adjust to end
    * @param {Boolean} [toTime=false] -- Adjust to time periods (timeIntervals)
    * @param {Number} [timeIntervals=60] -- Time period to adjust (in minutes)
    * @return {Date|Number|moment|msDateStr} adjustedDate -- Date in the input format
    */
-function tuneDateValue(origDate, toEnd, toTime, timeIntervals) {if (timeIntervals === void 0) {timeIntervals = 60;}
+function adjustDateValue(origDate, toEnd, toTime, timeIntervals) {if (timeIntervals === void 0) {timeIntervals = 60;}
   var dateType = detectDateValueType(origDate);
   var date;
   if (dateType === 'object') {// Just clone date object
@@ -46520,11 +46564,11 @@ function tuneDateValue(origDate, toEnd, toTime, timeIntervals) {if (timeInterval
 }
 
 function startOfTheDay(date) {
-  return tuneDateValue(date);
+  return adjustDateValue(date);
 }
 
 function endOfTheDay(date) {
-  return tuneDateValue(date, true);
+  return adjustDateValue(date, true);
 }
 
 /***/ }),