+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+- 2021.01.29, 00:29 -- FormItemHOC: Manually call focusIn/focusOut event handlers on focus/blur event handlers, added `framed` modifier. FormGroup: Styles for nested items/groups. FormButton (in `config/css`): set inner height to outer height due to usage of `box-shadow` instead `border` in styles.
+- 2021.01.29, 00:12 -- FormDateRange: Pass modal `open` state (as `checked`), minor changes (dates preparation in demo fixture).
+- 2021.01.29, 00:10 -- FormSelect, Menu, MenuItem, FormSelect: PropTypes updated, use `selectable`, `selected` parameters for multiple choices (consider `checkable` & `checked` for boolean-like elements). FormSelect: Pass open & checked state to popup control element, handle open state from `ModalPopup`, tune styles for fullWidth control element.
+- 2021.01.28, 23:52 -- utils/dates: Minor changes (dayTicks, weekTicks exported constants).
+- 2021.01.28, 23:50 -- FormButton: tabIndex=0, default tag is `button`, focusable param for FormItem, inputId attribute, correct blur event processing.
 - 2021.01.28, 20:09 -- v.0.2.8: Fixed PropTypes for DateTimeSelector, FormDateRange, FormDateTime, FormGroup (added props `wrap`, `noSpace`), FormItemHOC, FormRadio. FormDateRange: Use event handlers `onStartDateChange`, `onEndDateChange`. Added `utils.dates` methods: `startOfTheDay`, `endOfTheDay` (fixed exports in `utils`). Minor maintenance tools fixups.
 - 2021.01.26, 21:01 -- v.0.2.7.
 - 2021.01.26, 20:37 -- PD-3546: DateTimeSelector, FormDateRange: Support for date/time range selection (done).
@@ -78,5 +83,5 @@
 - 2020.10.05, 20:38 -- v.0.0.15: Extracted `demoSupport.jsx` routines, fixture selection from url query (parameter `fixture`), added avaialble fixtures list if fixture to display is not specified, added DemoWrapper & demoTitle exportable ficture parameters for demo rendering.
 
 <!--
- @changed 2021.01.28, 20:18
+ @changed 2021.01.29, 00:29
 -->